refactor: rename devices panel components by UI role
