--- src/components/device/DeviceList.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import OtherDevice from './OtherDevice'
-
-const DeviceList: React.FC = () => {
-  return (
-    <>
-      <OtherDevice />
-    </>
-  )
-}
-
-export default DeviceList
