+++ src/pages/__tests__/DevicesPage.test.tsx
@@ -3,9 +3,9 @@ import { describe, expect, it, vi } from 'vitest'
 import DevicesPage from '@/pages/DevicesPage'
 
 vi.mock('@/components', () => ({
-  DeviceList: () => (
+  PairedDevicesPanel: () => (
     <div>
-      <div data-testid="device-list">DeviceList</div>
+      <div data-testid="paired-devices-panel">PairedDevicesPanel</div>
     </div>
   ),
   PairingDialog: () => <div data-testid="pairing-dialog-mock">PairingDialog</div>,
@@ -15,7 +15,7 @@ describe('DevicesPage', () => {
   it('renders device list and does not render legacy sections', () => {
     render(<DevicesPage />)
 
-    expect(screen.getByTestId('device-list')).toBeInTheDocument()
+    expect(screen.getByTestId('paired-devices-panel')).toBeInTheDocument()
     expect(screen.queryByText('Device Management')).not.toBeInTheDocument()
     expect(screen.queryByText('Pairing Requests')).not.toBeInTheDocument()
     expect(screen.queryByText('当前设备')).not.toBeInTheDocument()
